fix job Generate client crt
according to documentation :

shell – Execute shell commands on targets
It is almost exactly like the command module but runs the command through a shell (/bin/sh) on the remote node.

and:

command – Execute commands on targets
The command(s) will not be processed through the shell, so variables like $HOME and operations like "<", ">", "|", ";" and "&" will not work. Use the shell module if you need these features.